Fixed a few more system dependent type errors (size_t vs unsigned vs long unsigned).

+++ drizzled/handler.cc
@@ -2938,8 +2938,8 @@ bool DsMrr_impl::get_disk_sweep_mrr_cost(uint32_t keynr, ha_rows rows, uint32_t 
   else
   {
     cost->zero();
-    *buffer_size= max((ulong)*buffer_size,
-                      (size_t)(1.2*rows_in_last_step) * elem_size +
+    *buffer_size= max(*buffer_size,
+                      (uint32_t)(1.2*rows_in_last_step) * elem_size +
                       h->ref_length + table->key_info[keynr].key_length);
   }
